Issue #5 Merged registry and factory
