+++ src/main/java/org/zalando/failsafeactuator/service/CircuitBreakerRegistry.java
@@ -1,18 +1,22 @@
 /**
- * The MIT License (MIT)
- * Copyright (c) 2016 Zalando SE
+ * The MIT License (MIT) Copyright (c) 2016 Zalando SE
  *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+ * <p>Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
+ * of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
  *
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+ * <p>The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
  *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ * <p>THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+ * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 package org.zalando.failsafeactuator.service;
 
 import net.jodah.failsafe.CircuitBreaker;
 
 import org.springframework.util.Assert;
+import org.zalando.failsafeactuator.endpoint.FailsafeEndpoint;
 
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
@@ -36,8 +40,8 @@ public class CircuitBreakerRegistry {
    * @param name Which is used to identify the CircuitBreaker
    */
   void registerCircuitBreaker(final CircuitBreaker breaker, final String name) {
-    Assert.hasText(name, "Name for circuitbreaker needs to be set");
-    Assert.notNull(breaker, "Circuitbreaker to add, can't be null");
+    Assert.hasText(name, "Name for circuit breaker needs to be set");
+    Assert.notNull(breaker, "Circuit breaker to add, can't be null");
 
     final CircuitBreaker replaced = concurrentBreakerMap.put(name, breaker);
     Assert.isNull(replaced, "There was an Circuit-Breaker registered already with name : " + name);
@@ -55,8 +59,7 @@ public class CircuitBreakerRegistry {
   }
 
   /**
-   * Returns the {@link CircuitBreaker} for the given name or <code>null</code> if no breaker with
-   * that name was registered before.
+   * Returns the {@link CircuitBreaker} for the given name or <code>null</code> if no breaker with that name was registered before.
    *
    * @param name of the CircuitBreaker to get
    * @return the found CircuitBreaker or <code>null</code>
@@ -73,6 +76,32 @@ public class CircuitBreakerRegistry {
    */
   public Map<String, CircuitBreaker> getConcurrentBreakerMap() {
     return this.concurrentBreakerMap;
+  }
+
+  /**
+   * Creates a new {@link CircuitBreaker} which can be used in regular way and is registered in the SpringContext.
+   *
+   * @param identifier which will be shown in the output of the {@link FailsafeEndpoint}.
+   * @return new Instance of a {@link CircuitBreaker}
+   */
+  private CircuitBreaker createCircuitBreaker(final String identifier) {
+    final CircuitBreaker breaker = new CircuitBreaker();
+    registerCircuitBreaker(breaker, identifier);
+    return breaker;
+  }
+
+  /**
+   * Returns the registered {@link CircuitBreaker} for the given identifier or creates a new one.
+   *
+   * @param identifier which will be shown in the output of the {@link FailsafeEndpoint}.
+   * @return an already registered or a new Instance of a {@link CircuitBreaker}
+   */
+  public CircuitBreaker getOrCreate(final String identifier) {
+    if (contains(identifier)) {
+      return get(identifier);
+    } else {
+      return createCircuitBreaker(identifier);
+    }
   }
 
   @PreDestroy
