+++ src/test/java/org/zalando/failsafeactuator/endpoint/EndpointTest.java
@@ -1,12 +1,15 @@
 /**
- * The MIT License (MIT)
- * Copyright (c) 2016 Zalando SE
+ * The MIT License (MIT) Copyright (c) 2016 Zalando SE
  *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+ * <p>Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
+ * of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
  *
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+ * <p>The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
  *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ * <p>THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+ * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 package org.zalando.failsafeactuator.endpoint;
 
@@ -24,8 +27,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.zalando.failsafeactuator.endpoint.domain.CircuitBreakerState;
-import org.zalando.failsafeactuator.service.CircuitBreakerFactory;
-import org.zalando.failsafeactuator.service.CircuitBreakerRegistry;
+import org.zalando.failsafeactuator.service.FailsafeBreaker;
 
 import java.util.Arrays;
 
@@ -50,19 +52,11 @@ public class EndpointTest {
   private TestRestTemplate restTemplate;
 
   @Autowired
-  private CircuitBreakerFactory factory;
-
-  @Autowired
-  private CircuitBreakerRegistry registry;
-
-  @Before
-  public void setup() {
-    registry.getConcurrentBreakerMap().clear();
-  }
+  @FailsafeBreaker(BREAKER_NAME)
+  private CircuitBreaker breaker;
 
   @Test
   public void endpointTest() {
-    final CircuitBreaker breaker = factory.createCircuitBreaker(BREAKER_NAME);
     CircuitBreakerState state = fetchCircuitBreaker();
     assertTrue(state.isClosed());
     assertFalse(state.isOpen());
@@ -82,9 +76,7 @@ public class EndpointTest {
     assertFalse(state.isOpen());
     assertTrue(state.isHalfOpen());
     assertEquals(BREAKER_NAME, state.getName());
-
   }
-
 
   private CircuitBreakerState fetchCircuitBreaker() {
     final ResponseEntity<CircuitBreakerState[]> result = restTemplate.getForEntity(String.format(FAILSAFE_URL, port), CircuitBreakerState[].class);
